feat(bot/core): add headers.py file to store headers for requests
This commit adds a new file `headers.py` to the `bot/core` directory. This file contains a dictionary of headers that will be used for making requests to the CEX.IO API. This will help to ensure that the requests are made with the correct headers and that the bot can access the API correctly.

Signed-off-by: troublescope <tomiprs25@gmail.com>